Move manual tests to tests/ directory to fix error when package is pip installed
