+++ pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
 norecursedirs = tests/data
+markers =
+    slow: mark a test as a slow test or requiring manual intervention (such as opening a GUI window).
