+++ treehouse/dataflow/reaching_def.py
@@ -1,4 +1,3 @@
-from tests.utils import draw
 from treehouse.tree_sitter_utils import c_parser
 from treehouse.ast_creator import ASTCreator
 from treehouse.cfg_creator import CFGCreator
@@ -87,30 +86,3 @@ class ReachingDefinitionSolver(DataflowSolver):
                     return self.id2def[i]
         else:
             return set()
-
-
-def test():
-    code = """int main()
-    {
-        int x = 0;
-        if (true) {
-            x += 5;
-        }
-        printf("%d\\n", x);
-        x = 10;
-        return x;
-    }
-    """
-    tree = c_parser.parse(bytes(code, "utf8"))
-    ast = ASTCreator.make_ast(tree.root_node)
-    cfg = CFGCreator.make_cfg(ast)
-    solver = ReachingDefinitionSolver(cfg, verbose=1)
-    solution_in, solution_out = solver.solve()
-    draw(
-        cfg,
-        {
-            n: sorted(set(map(solver.def2code.__getitem__, b)))
-            for n, b in solution_out.items()
-        },
-    )
-    print(solution_out)
